requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
pydantic
python-hcl2
pyyaml
types-pyyaml
